% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abundance}
\alias{abundance}
\title{Microbial abundance time series}
\format{
A data frame with 46046 rows and 19 variables:
\describe{
  \item{isolate.id}{name of a given phytoplankton isolate}
  \item{experiment.ID}{unique identifier combining isolate, temperature, and dilution number}
  \item{dilution}{dilution number, ranging from 1-3. Populations were periodically diluted to provide new nutrients for growth}
  \item{replicate}{experimental replicate (4 replicates per temperature and isolate)}
  \item{plate.file}{name of data file produced by plate reader, containing fluorescence readings}
  \item{date}{date of measurement, yyyy-mm-dd format}
  \item{time}{time of day of measurement, hh:mm}
  \item{media}{growth media use for culturing populations}
  \item{well}{location of specific population in the 24-well plate}
  \item{fluorescence}{measured fluorescence of population, a standard proxy for abundance}
  \item{temperature}{temperature, in Celsius}
  \item{dtime}{time since first measure of population abundance, in hours}
  \item{log.fluor}{natural log of fluorescence}
  \item{genus}{genus of phytoplankton population}
  \item{species}{species epithet of phytoplankton population}
  \item{group}{functional group phytoplankton species belongs to}
  \item{source}{location where phytoplankton species were isolated, NB is Narragansett Bay}
  \item{plate.temperature}{temperature of well plate during fluorescence readings}
  \item{id}{unique identifier for single experimental phytoplankton populations}
}
}
\source{
Elena Litchman and Tatiana Severin, W.K. Kellogg Biological Station, Michigan State University
}
\usage{
data(abundance)
}
\description{
A data set tracking the abundances of experimental phytoplankton populations over time. Fluorescence 
is used as a proxy for numerical abundance/biomass.
}
\keyword{datasets}
