% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{augment.growth_fit}
\alias{augment.growth_fit}
\title{Augment method for growth model fit}
\usage{
\method{augment}{growth_fit}(object, ...)
}
\arguments{
\item{object}{Object of class growth_fit}

\item{\dots}{Additional arguments (not used)}
}
\description{
Provides observation-level predictions and residuals based on a model fit
}
