% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{decurve}
\alias{decurve}
\title{Double exponential model}
\usage{
decurve(temperature, topt, b1, b2, d0, d2)
}
\arguments{
\item{temperature}{Temperature(s) to calculate growth rate at}

\item{topt}{Optimum temperature (where growth rate is highest)}

\item{b1}{Birth rate at 0 Celsius}

\item{b2}{Temperature scaling of birth rate, > 0}

\item{d0}{Temperature-independent death rate}

\item{d2}{Temperature scaling of death rate, > 0 (and d2 > b2)}
}
\value{
Growth rate at one or more temperatures
}
\description{
Re-parameterized from Thomas et al. 2016 to depend explicitly on Topt
}
