% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{decurve2}
\alias{decurve2}
\title{Double exponential model, phi version}
\usage{
decurve2(temp, topt, phi, b2, d0, d2)
}
\arguments{
\item{temp}{Temperature(s) to calculate growth rate at}

\item{topt}{Optimum temperature (where growth rate is highest)}

\item{phi}{Birth rate at 0 Celsius}

\item{b2}{Temperature scaling of birth rate, > 0}

\item{d0}{Temperature-independent death rate}

\item{d2}{Temperature scaling of death rate, > 0 (and d2 > b2)}
}
\value{
Growth rate at one or more temperatures
}
\description{
Alternate parameterization for Topt and phi (see mathematica notebook)
}

