% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{derive.satdecay.stats}
\alias{derive.satdecay.stats}
\title{Derive set of additional traits for satdecayode model}
\usage{
derive.satdecay.stats(cfs)
}
\arguments{
\item{cfs}{Coefficients of satdecay ODE model, e.g. from get.gr.satdecay.ode(), including r0}
}
\value{
List containing tmax (timing of peak abundance) and nmax (ln(abundance) at tmax)
}
\description{
Given a set of estimated coefficients for a satdecay ODE model, use numerical
methods in Julia to calculate additional values of interest, including peak ln(abundance).
}
