% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{detect}
\alias{detect}
\title{Detect model failure}
\usage{
detect(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE/FALSE depending on whether or not 'try-error' is listed as (one) of the classes of object x.
}
\description{
Helper function used to determine whether an attempt at an mle2 fit failed
}
