% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{.julia_initialized}
\alias{.julia_initialized}
\title{Start assuming Julia is not initialized}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.julia_initialized
}
\description{
Start assuming Julia is not initialized
}
\keyword{datasets}
