% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{droopcurve}
\alias{droopcurve}
\title{Droop extension of double exponential model}
\usage{
droopcurve(temperature, b1, b2, d0, d1, d2, X)
}
\arguments{
\item{temperature}{Temperature(s) to calculate growth rate at}

\item{b1}{Birth rate at 0 Celsius}

\item{b2}{Temperature scaling of birth rate, > 0}

\item{d0}{Temperature-independent death rate}

\item{d2}{Temperature scaling of death rate, > 0 (and d2 > b2)}

\item{X}{ratio of Qmin to Vmax}
}
\value{
Growth rate at one or more temperatures
}
\description{
Developed by David Anderson, emerges from considering effects of nutrient limitation
in a quota-based growth model (Droop model), when it is assumed that nutrient quota
equillibrates faster than population dynamics.
}
