% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ensure_julia}
\alias{ensure_julia}
\title{Make sure Julia is available if calling a function that needs it}
\usage{
ensure_julia()
}
\description{
Make sure Julia is available if calling a function that needs it
}
