% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_monod_data}
\alias{example_monod_data}
\title{Effect of phosphate on phytoplankton growth rates}
\format{
A data frame with 40 rows and 2 variables:
\describe{
  \item{phosphate}{phosphate concentration, in umol/L}
  \item{growth.rate}{estimated growth rate}
}
}
\source{
Elena Litchman, Allyson Hutchins, Tatiana Severin, W.K. Kellogg Biological Station, Michigan State University
}
\usage{
data(example_monod_data)
}
\description{
A data set containing estimated growth rates of one phytoplankton isolate across a range
of phosphate concentrations. Growth rates were estimated from abundance time series (abundance data set) 
using methods in the growthTools package.
}
\keyword{datasets}
