% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_TPC_data}
\alias{example_TPC_data}
\title{Effect of temperature on phytoplankton growth rates}
\format{
A data frame with 160 rows and 8 variables:
\describe{
  \item{experiment.ID}{unique identifier combining isolate, temperature, and dilution number}
  \item{isolate.id}{name of a given phytoplankton isolate}
  \item{temperature}{temperature, in Celsius}
  \item{dilution}{dilution number, ranging from 1-3}
  \item{replicate}{experimental replicate, 4 replicates per temperature and isolate}
  \item{mu}{estimated growth rate}
  \item{best.model}{time series model generating estimated growth rate}
}
}
\source{
Elena Litchman and Tatiana Severin, W.K. Kellogg Biological Station, Michigan State University
}
\usage{
data(example_TPC_data)
}
\description{
A data set containing estimated growth rates of two phytoplankton isolates across a range
of temperatures. Growth rates were estimated from abundance time series (abundance data set) 
using methods in the growthTools package.
}
\keyword{datasets}
