% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fd2.central}
\alias{fd2.central}
\title{Approximate 2nd derivative (finite difference)}
\usage{
fd2.central(fx, h)
}
\arguments{
\item{fx}{A vector of 3 function values, f(x-h), f(x), f(x+h)}

\item{h}{The step size used in the finite difference calculation}
}
\description{
Approximate 2nd derivative (finite difference)
}
