% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{fit_model.linear_model}
\alias{fit_model.linear_model}
\title{Fitting methods for growth models}
\usage{
\method{fit_model}{linear_model}(model, x, y, ...)
}
\arguments{
\item{model}{Growth model type}

\item{x}{Time variable}

\item{y}{ln(abundance)}

\item{\dots}{Additional arguments passed to fitting function (not used?)}
}
\description{
Suite of fitting methods applied to different growth models in order to fit
variation in ln(abundance) over time and estimate population exponential growth
rate. Input follows a common structure, including model type, and redirects
to specific get.gr() style function. These could probably subsume the get.gr()
functions directly, instead of simply providing an alias, which would simplify 
things
}
