% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{get.decurve.tpc}
\alias{get.decurve.tpc}
\title{Fit Double Exponential curve to growth rate vs. temperature data}
\usage{
get.decurve.tpc(temp, mu, method = "grid.mle2",
  start.method = "general.grid", plotQ = F, conf.bandQ = T, fpath = NA,
  id = NA, ...)
}
\arguments{
\item{temp}{Temperature}

\item{mu}{Exponential growth rate}

\item{plotQ}{Should regression be visualized?}

\item{conf.bandQ}{Should we calculate a confidence band around the regression? logical.}

\item{fpath}{If visual requested, and valid file path provided here, plot will be saved as a .pdf file. Default is NA.}

\item{id}{Character string providing any information ID'ing the specifc curve being fit; used to label plots, if any are requested. Default is NA.}

\item{fit.method}{Specify which fitting algorithm to use, 'mle2' or 'grid.mle2'}
}
\description{
Fit Double Exponential curve to growth rate vs. temperature data
}

