% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{get.droopcurve.tpc}
\alias{get.droopcurve.tpc}
\title{Fit Droop-based Double Exponential curve to growth rate vs. temperature data}
\usage{
get.droopcurve.tpc(
  temperature,
  mu,
  method = "grid.mle2",
  start.method = "general.grid",
  suppress.grid.mle2.warnings = TRUE,
  ...
)
}
\arguments{
\item{temperature}{Temperature}

\item{mu}{Exponential growth rate}

\item{method}{Specify which fitting algorithm to use, 'mle2' or 'grid.mle2'}

\item{start.method}{Specify method for generating starting grid for 'grid.mle2' option}

\item{suppress.grid.mle2.warnings}{logical; should warnings arising from grid.mle2 invocation be suppressed (TRUE), or displayed (FALSE)? Default is TRUE.}

\item{...}{Additional arguments passed to grid.mle2 (e.g., control=list(maxit=2000))}
}
\description{
Fit Droop-based Double Exponential curve to growth rate vs. temperature data
}
