% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_tools.R
\name{get.ep}
\alias{get.ep}
\title{Fit Eiler Peeters curve to growth rate vs. light data}
\usage{
get.ep(light, mu, method = "mle2", ...)
}
\arguments{
\item{light}{Light level}

\item{mu}{Exponential growth rate}

\item{method}{Specify which fitting algorithm to use, 'mle2' or 'grid.mle2'}

\item{...}{Additional arguments passed to grid.mle2 (e.g., control=list(maxit=2000))}
}
\description{
Note: this function currently does not use grid.mle2, as the regressions are usually pretty stable; add this 
feature later.
}
\examples{

}
