% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.gr.linear}
\alias{get.gr.linear}
\title{Extract exponential growth rate assuming exponential growth}
\usage{
get.gr.linear(x, y, plotQ = F, fpath = NA, id = "")
}
\arguments{
\item{x}{Time steps}

\item{y}{ln(abundance)}

\item{plotQ}{logical; should the fit be plotted?}

\item{fpath}{character; path specifying where plot should be saved, if generated}

\item{id}{Label corresponding to the population/strain/species of interest}
}
\value{
This function returns a linear model regressing ln(abundance) on time
}
\description{
This function fits a linear model to ln(abundance) data.
}
