% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.gr.satdecay.ode}
\alias{get.gr.satdecay.ode}
\title{Extract exponential growth rate assuming exponential growth that saturates and 
then decays using full ODE fitting}
\usage{
get.gr.satdecay.ode(x, y, plotQ = F, fpath = NA, id = "")
}
\arguments{
\item{x}{Time steps}

\item{y}{ln(abundance)}
}
\value{
This function returns an mle2 regression model
}
\description{
This function fits ODE model to ln(abundance) data, with 
the assumption that abundances increase linearly at first, as resource is 
depleted in the background. If resources are perfectly recycled into the 
resource pool, abundances will saturate. But, imperfect recycling leads to
subsequent decay in abundance as cells die, and recalcitrant nutrients 
accumulate. Under particular circumstances, this dynamic is well approximated
by a piecewise linear function; see get.gr.satdecay(). Currently, this 
functionality only works if R is connected to Julia, which provides fast
ODE solutions
}
