% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{get.tlim}
\alias{get.tlim}
\title{Calculate thermal niche limits}
\usage{
get.tlim(topt, w, b, type = "tmin")
}
\arguments{
\item{topt}{Competitive optimum temperature}

\item{w}{Thermal niche width}

\item{b}{Exponential scaling}

\item{type}{One of either 'tmin' or 'tmax'; default is 'tmin'}
}
\value{
Either the estimate of Tmin or Tmax corresponding to the provided Norberg parameters.
}
\description{
Note: check whether this operates on nbcurve() or nbcurve2() parameters!!!
}
