% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{get.topt}
\alias{get.topt}
\title{Get optimum temperature from nbcurve.legacy()}
\usage{
get.topt(copt, w, b)
}
\arguments{
\item{copt}{Competitive optimum temperature}

\item{w}{Thermal niche width}

\item{b}{Exponential scaling}
}
\value{
Optimum temperature
}
\description{
Given the competitive optimum temperature and other legacy Norberg parameters, 
calculate the true optimum temperature.
}
