% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{glance.growth_fit}
\alias{glance.growth_fit}
\title{Glance method for single growth model fit}
\usage{
\method{glance}{growth_fit}(x, ...)
}
\arguments{
\item{\dots}{Additional arguments (not used)}

\item{object}{Object of class growth_fit}
}
\description{
Provides observation-level predictions and residuals based on a model fit
}
