% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthTools.R
\docType{package}
\name{growthTools-package}
\alias{growthTools-package}
\title{define as a package (replaces former '@docType package')}
\description{
Developed for applications to NSF Dimensions of Biodiversity grant.
}
