% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{localslope}
\alias{localslope}
\title{Local Slope function}
\usage{
localslope(d)
}
\arguments{
\item{d}{A data frame containing two columns, x and y}
}
\value{
Slope of the linear regression
}
\description{
Helper function to calculate and extract the slope of a basic linear regression 
relating y to x; the resulting value is used to obtain a reasonable starting guess
for the slopes of the piecewise linear functions in \code{lag}, \code{sat}, and
 \code{lagsat}
}

