% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_tools.R
\name{logLik.npc}
\alias{logLik.npc}
\title{logLik extraction method for NPC objects}
\usage{
logLik.npc(x, ...)
}
\arguments{
\item{x}{Object of class npc}

\item{\dots}{Additional arguments (not used)}
}
\description{
logLik extraction method for NPC objects
}
