% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools_S3.R
\name{logLik.tpc}
\alias{logLik.tpc}
\title{logLik extraction method for TPC objects}
\usage{
logLik.tpc(x, ...)
}
\arguments{
\item{x}{Object of class tpc}

\item{\dots}{Additional arguments (not used)}
}
\description{
logLik extraction method for TPC objects
}
