% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_tools.R, R/nutrient_tools.R
\name{monod_curve_type2}
\alias{monod_curve_type2}
\title{Monod curve with loss term equation}
\usage{
monod_curve_type2(x, umax, k, z, log.pars = FALSE)

monod_curve_type2(x, umax, k, z, log.pars = FALSE)
}
\arguments{
\item{x}{nutrients}

\item{umax}{maximum growth rate}

\item{k}{half saturation constant}

\item{z}{ln(intercept)}

\item{log.pars}{logical, are parameters specified on natural log scale}
}
\value{
Predicted exponential growth rate at nutrient concentration x

Predicted exponential growth rate at nutrient concentration x
}
\description{
Details here.

Details here.
}
