% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{nbcurve}
\alias{nbcurve}
\title{Re-parameterized Norberg equation}
\usage{
nbcurve(x, topt, w, a, b)
}
\arguments{
\item{x}{Temperature}

\item{topt}{Optimum temperature}

\item{w}{Thermal niche width}

\item{a}{Affects the y-intercept}

\item{b}{Exponential scaling}
}
\value{
Predicted exponential growth rate at temperature x
}
\description{
One standard thermal performance curve model comes from Norberg et al. 2004, where growth rate varies
with temperature as the product of an exponential and a quadratic function. A significant limitation of 
this formulation is that the 'optimum' temperature parameter in Norberg et al. 2004 does not actually 
correspond exactly to the temperature at which growth rate is maximized (see Thomas et al. 2012, etc.).
However, it is possible to re-parameterize the Norberg equation to depend explicitly on a parameter Topt 
that does correspond as expected to maximum growth rate (e.g., Baker et al. 2016). We use the latter form
of the Norberg equation in this package. Users wishing to access the original equation can see `?nbcurve.legacy`
}
\details{
Mathematically, the re-parameterized Norberg equation is given by:
\deqn{\mu = e^{a+bT} (1 - ((\phi/(b w))^2)}
where \eqn{\phi = -2 - 2 b Topt + 2 b T + (4 + (b w)^2)^(1/2)}
}
