% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{new_growth_rate_result}
\alias{new_growth_rate_result}
\title{Growth rate result class creation}
\usage{
new_growth_rate_result(models, successful, best, ictab)
}
\arguments{
\item{models}{Set of growth rate models employed}

\item{successful}{Which model(s) fit successfully}

\item{best}{Growth rate model selected as the best model}

\item{ictab}{Model comparison results (information criteria table)}
}
\description{
Growth rate result class creation
}
