% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_tools.R, R/nutrient_tools.R
\name{new_npc}
\alias{new_npc}
\title{Nutrient performance curve (NPC) class creation}
\usage{
new_npc()

new_npc()
}
\description{
Nutrient performance curve (NPC) class creation

Nutrient performance curve (NPC) class creation
}
