% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nobs.mle2}
\alias{nobs.mle2}
\title{Extract number of observations from mle2 object}
\usage{
nobs.mle2(object)
}
\description{
By default, mle2 objects from bbmle produced by fitting a model to data without
an explicit formula (e.g., using a black-boxed NLL calculator function instead)
do not provide the number of observations underlying the fit. This extends the 
nobs() generic with a specific method for mle2 objects.
}
