% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{plot.decurve}
\alias{plot.decurve}
\title{Plotting function for single Double Exponential curve}
\usage{
plot.decurve(fit.info, plot.ci = TRUE, plot.obs = TRUE, xlim = c(-2,
  40), ylim = c(-0.2, 5))
}
\arguments{
\item{fit.info}{The result of a single TPC curve fit from get.decurve.tpc}

\item{plot.ci}{logical, should the resulting plot include 95% confidence bands}

\item{plot.obs}{logical, should resulting plot include raw data}

\item{xlim}{x-axis range (temperature)}

\item{ylim}{y-axis range (adjusts internally to -0.2 to slightly above umax+CI)}
}
\description{
Plotting function for single Double Exponential curve
}
