% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{plot.growth_fit}
\alias{plot.growth_fit}
\title{Generic plotting method for growth_fit object}
\usage{
\method{plot}{growth_fit}(object, main = NULL, savepath = NULL, ngrid = 200, ...)
}
\arguments{
\item{object}{Object of class growth_fit}

\item{\dots}{Additional arguments (not used)}

\item{newdata}{New data (if any) to use for making predictions}
}
\description{
Generic plotting method for growth_fit object
}
