% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_tools.R
\name{plot.npc}
\alias{plot.npc}
\title{Plotting function for single nutrient performance curve}
\usage{
\method{plot}{npc}(
  x,
  plot_ci = TRUE,
  plot_obs = TRUE,
  xlim = c(0, 100),
  ylim = c(-0.2, 5),
  main = NA,
  fpath = NA,
  ...
)
}
\arguments{
\item{x}{The result of a single NPC curve fit from get.monod or similar}

\item{plot_ci}{logical, should the resulting plot include 95\% confidence bands}

\item{plot_obs}{logical, should resulting plot include raw data}

\item{xlim}{x-axis range (nutrients)}

\item{ylim}{y-axis range (adjusts internally from -0.2 to slightly above umax+CI)}

\item{main}{Character string providing plot title (usually id info for the plotted curve)}

\item{fpath}{If visual requested, and valid file path provided, save plot as a .pdf file}

\item{\dots}{Additional arguments (not used)}
}
\description{
Plotting function for single nutrient performance curve
}
