% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{plot_model}
\alias{plot_model}
\title{Plotting methods for growth models}
\usage{
plot_model(model, fit, x, y, ...)
}
\arguments{
\item{model}{Growth model type}

\item{fit}{Actual fit of growth model}

\item{x}{Time variable}

\item{y}{ln(abundance)}

\item{\dots}{Additional arguments (not used)}
}
\description{
Generic for a suite of methods used to plot the fits of different growth 
models. Specific methods have yet to be written.
}
