% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{predict.decurve}
\alias{predict.decurve}
\title{Predict values from decurve fit}
\usage{
predict.decurve(
  fit.info,
  newdata = data.frame(temperature = seq(-2, 40, 0.1)),
  se.fit = FALSE
)
}
\arguments{
\item{fit.info}{The result of a single TPC curve fit from get.decurve.tpc}

\item{newdata}{A new data frame, containing a sequence of `temperature` values at which model predictions should be made; defaults to (-2,40)}

\item{se.fit}{logical; should standard error values be returned?}
}
\description{
Predict values from decurve fit
}
