% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{predict.growth_rate_result}
\alias{predict.growth_rate_result}
\title{Predict method for best model in the suite of growth rate results}
\usage{
predict.growth_rate_result(object, ...)
}
\arguments{
\item{object}{Object of class growth_rate_result}

\item{\dots}{Additional arguments (not used)}
}
\description{
Predict method for best model in the suite of growth rate results
}
