% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_tools.R
\name{predict.lpc}
\alias{predict.lpc}
\title{Predict values from an LPC fit}
\usage{
\method{predict}{lpc}(object, newdata, se.fit = FALSE, ...)
}
\arguments{
\item{object}{The result of a single LPC curve fit, an object of class lpc}

\item{newdata}{A new data frame, containing a sequence of `light` values at which model predictions should be made; defaults to (0,400)}

\item{se.fit}{logical; should standard error values be returned?}

\item{\dots}{Additional arguments (not used)}
}
\description{
Predict values from an LPC fit
}
