% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{predict.nbcurve}
\alias{predict.nbcurve}
\title{Predict values from nbcurve fit}
\usage{
predict.nbcurve(fit.info, newdata = data.frame(temperature = seq(-2, 40,
  0.1)), se.fit = FALSE)
}
\arguments{
\item{fit.info}{The result of a single TPC curve fit from get.nbcurve.tpc}

\item{newdata}{A new data frame, containing a sequence of `temperature` values at which model predictions should be made; defaults to (-2,40)}

\item{se.fit}{logical; should standard error values be returned?}
}
\description{
Predict values from nbcurve fit
}
