% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_tools.R
\name{predict.npc}
\alias{predict.npc}
\title{Predict values from an NPC fit}
\usage{
\method{predict}{npc}(object, newdata, se.fit = FALSE, ...)
}
\arguments{
\item{object}{The result of a single NPC curve fit from get.monod, an object of class npc}

\item{newdata}{A new data frame, containing a sequence of `nutrient` values at which model predictions should be made; defaults to (0,100)}

\item{se.fit}{logical; should standard error values be returned?}

\item{\dots}{Additional arguments (not used)}
}
\description{
Predict values from an NPC fit
}
