% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{prepare_growth_data}
\alias{prepare_growth_data}
\title{Check at prepare abundance time series for analysis}
\usage{
prepare_growth_data(x, y, zero.time = TRUE)
}
\arguments{
\item{x}{time values}

\item{y}{ln(abundance) values}

\item{zero.time}{if TRUE, shift time axis so that each time series starts at time = 0}
}
\description{
Check at prepare abundance time series for analysis
}
