% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{print.growth_fit}
\alias{print.growth_fit}
\title{Print method for growth_fit object}
\usage{
\method{print}{growth_fit}(object, ...)
}
\arguments{
\item{object}{Object of class growth_fit}

\item{\dots}{Additional arguments (not used)}
}
\description{
Print method for growth_fit object
}
