% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{print.growth_rate_result}
\alias{print.growth_rate_result}
\title{Print method for best model in the suite of growth rate results}
\usage{
\method{print}{growth_rate_result}(object, ...)
}
\arguments{
\item{object}{Object of class growth_rate_result}

\item{\dots}{Additional arguments (not used)}
}
\description{
Print method for best model in the suite of growth rate results
}
