% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_tools.R
\name{print.lpc}
\alias{print.lpc}
\title{Print method for LPC objects}
\usage{
\method{print}{lpc}(x, ...)
}
\arguments{
\item{x}{Object of class lpc}

\item{\dots}{Additional arguments (not used)}
}
\description{
Print method for LPC objects
}
