% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools_S3.R
\name{print.tpc}
\alias{print.tpc}
\title{Print method for TPC objects}
\usage{
\method{print}{tpc}(x, ...)
}
\arguments{
\item{x}{Object of class tpc}

\item{\dots}{Additional arguments (not used)}
}
\description{
Print method for TPC objects
}
