% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{r0.global}
\alias{r0.global}
\title{Define global parameters}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
r0.global
}
\description{
Define global parameters
}
\keyword{datasets}
