% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{run_growth_model}
\alias{run_growth_model}
\title{Fit and evaluate specific growth rate model}
\usage{
run_growth_model(model, x, y, min.exp.obs = 3, internal.r2.cutoff = 0)
}
\arguments{
\item{model}{Desired model type}

\item{x}{Time steps}

\item{y}{ln(abundance)}

\item{min.exp.obs}{control parameter specifying the minimum number of observations that must fall within the estimated exponential phase in order to consider lag, sat, lagsat, and flr models; defaults to 3.}

\item{internal.r2.cutoff}{control parameter specifying the R2 criteria that may be applied to drop fits where the number of observations in the exponential portion is equal to 3. The default value of zero permits all fits of 3 obs to be considered.}
}
\value{
This function returns a growth_fit object
}
\description{
This generic function fits a specific user-selected growth rate model to time
series data on ln(abundance). In addition to providing the final model fit, 
derived metrics are calculated after the model is fit (the evaluation step).
}
