% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{satdecay.ode.peak.time}
\alias{satdecay.ode.peak.time}
\title{Estimate timing of peak abundance (satdecay ODE model)}
\usage{
satdecay.ode.peak.time(cfs, r0 = r0.global, tmax = 100)
}
\arguments{
\item{cfs}{Coefficients of satdecay ODE model, e.g. from get.gr.satdecay.ode()}

\item{r0}{Initial resource concentration (set arbitrarily to r0.global by default)}

\item{tmax}{Maximum of time domain}
}
\description{
Given a set of estimated coefficients for a satdecay ODE model, use numerical
methods in Julia to calculate the timing of peak ln(abundance).
}
