% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{summary.growth_rate_result}
\alias{summary.growth_rate_result}
\title{Summary method for suite of successful growth rate results}
\usage{
\method{summary}{growth_rate_result}(object, ...)
}
\arguments{
\item{object}{Object of class growth_rate_result}

\item{\dots}{Additional arguments (not used)}
}
\description{
Summary method for suite of successful growth rate results
}
