% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_tools.R
\name{ep_curve}
\alias{ep_curve}
\title{Eiler-Peeters curve equation}
\usage{
ep_curve(x, umax, k, log.k = FALSE, log.umax = FALSE)
}
\arguments{
\item{x}{nutrients}

\item{umax}{maximum growth rate}

\item{k}{half saturation constant}

\item{log.k}{logical, is k specified on natural log scale}

\item{log.umax}{logical, is umax specified on natural log scale}
}
\value{
Predicted exponential growth rate at nutrient concentration x
}
\description{
Details here.
}
