% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{evaluate_model.linear_model}
\alias{evaluate_model.linear_model}
\alias{evaluate_model.lag_model}
\alias{evaluate_model.sat_model}
\alias{evaluate_model.flr_model}
\alias{evaluate_model.satdecay_model}
\alias{evaluate_model.lagsat_model}
\title{Evaluation methods for growth models}
\usage{
\method{evaluate_model}{linear_model}(model, fit, x, y, ...)

\method{evaluate_model}{lag_model}(model, fit, x, y, ...)

\method{evaluate_model}{sat_model}(model, fit, x, y, ...)

\method{evaluate_model}{flr_model}(model, fit, x, y, ...)

\method{evaluate_model}{satdecay_model}(model, fit, x, y, ...)

\method{evaluate_model}{lagsat_model}(model, fit, x, y, ...)
}
\arguments{
\item{model}{Growth model type}

\item{fit}{Actual fit of growth model}

\item{x}{Time variable}

\item{y}{ln(abundance)}

\item{\dots}{Additional arguments (not used)}
}
\description{
Suite of methods applied to different growth models in order to extract or 
calculate key metrics for evaluating the quality of model fits. This includes 
the exponential growth rate estimate itself, as well as measures for the 
fit in different portions of the model (e.g., initial exponential phase vs.
after saturating). Input follows a common structure, but internals vary by 
model type.
}
