% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{evaluate_model}
\alias{evaluate_model}
\alias{evaluate_model.linear_model}
\alias{evaluate_model.lag_model}
\alias{evaluate_model.sat_model}
\alias{evaluate_model.lagsat_model}
\alias{evaluate_model.flr_model}
\alias{evaluate_model.satdecay_model}
\alias{evaluate_model.satdecay_ode_model}
\title{Evaluation methods for growth models}
\usage{
evaluate_model(model, fit, x, y, ...)

\method{evaluate_model}{linear_model}(model, fit, x, y, ...)

\method{evaluate_model}{lag_model}(model, fit, x, y, ...)

\method{evaluate_model}{sat_model}(model, fit, x, y, ...)

\method{evaluate_model}{lagsat_model}(model, fit, x, y, ...)

\method{evaluate_model}{flr_model}(model, fit, x, y, ...)

\method{evaluate_model}{satdecay_model}(model, fit, x, y, ...)

\method{evaluate_model}{satdecay_ode_model}(model, fit, x, y, ...)
}
\arguments{
\item{model}{Growth model type}

\item{fit}{Actual fit of growth model}

\item{x}{Time variable}

\item{y}{ln(abundance)}

\item{\dots}{Additional arguments (not used)}
}
\description{
Generic for a suite of methods applied to different growth models to extract or 
calculate key metrics for evaluating the quality of model fits. This includes 
the exponential growth rate estimate itself, as well as measures for the 
fit in different portions of the model (e.g., initial exponential phase vs.
after saturating). Input follows a common structure, but internals vary by 
model type.
}
\details{
Note: when satdecayode fails to find an internal maximum abundance over the 
range of the data (either due to the data itself, or a poor fit), then function
satdecay.ode.peak.time() returns a tmax of NaN, and output of evaluate_model
changes. Most metrics are returned as NAs, and the slope provided is the decay
rate (if resulting curve is a linear decline). Further investigation is warranted
}
