% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_light_data}
\alias{example_light_data}
\title{Effect of light (irradiance) on phytoplankton growth rates}
\format{
A data frame with 36 rows and 3 variables:
\describe{
  \item{species}{species name}
  \item{light}{light level in photons m-2 s-1}
  \item{mu}{estimated growth rate}
}
}
\source{
Sam Fey, Tamara Layden, Reed College
}
\usage{
example_light_data
}
\description{
A data set containing estimated growth rates of one phytoplankton isolate across a range
of light levels.
}
\keyword{datasets}
