% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{fit_model}
\alias{fit_model}
\alias{fit_model.linear_model}
\alias{fit_model.lag_model}
\alias{fit_model.sat_model}
\alias{fit_model.flr_model}
\alias{fit_model.lagsat_model}
\alias{fit_model.satdecay_model}
\alias{fit_model.satdecay_ode_model}
\title{Fitting methods for growth models}
\usage{
fit_model(model, x, y, ...)

\method{fit_model}{linear_model}(model, x, y, ...)

\method{fit_model}{lag_model}(model, x, y, ...)

\method{fit_model}{sat_model}(model, x, y, ...)

\method{fit_model}{flr_model}(model, x, y, ...)

\method{fit_model}{lagsat_model}(model, x, y, ...)

\method{fit_model}{satdecay_model}(model, x, y, ...)

\method{fit_model}{satdecay_ode_model}(model, x, y, ...)
}
\arguments{
\item{model}{Growth model type}

\item{x}{Time variable}

\item{y}{ln(abundance)}

\item{\dots}{Additional arguments passed to fitting function (not used?)}
}
\description{
A suite of growth models is available which can be fit to time series of 
ln(abundance) over time and used to estimate population exponential growth
rate. The generic method is fit_model(), which dispatches to growth model
specific fitting functions, ultimately these are to specific get.gr() functions.
}
