% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{fit_model}
\alias{fit_model}
\title{Defining generics for fitting, evaluating, plotting}
\usage{
fit_model(model, x, y, ...)
}
\arguments{
\item{model}{Growth model type}

\item{x}{Time variable}

\item{y}{ln(abundance)}

\item{\dots}{Additional arguments passed to fitting function (not used?)}
}
\description{
Specifically, when these generics are called, dispatch will occur to the method
based on the first argument of the model object. This allows growth rate model
specific methods for fitting, plotting, etc.
}
