% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.gr.flr}
\alias{get.gr.flr}
\title{Extract exponential growth rate assuming exponential death that hits a floor}
\usage{
get.gr.flr(x, y, plotQ = F, fpath = NA, id = "")
}
\arguments{
\item{x}{Time steps}

\item{y}{ln(abundance)}

\item{plotQ}{logical; should the fit be plotted?}

\item{fpath}{character; path specifying where plot should be saved, if generated}

\item{id}{Label corresponding to the population/strain/species of interest}
}
\value{
This function returns a nonlinear least-squares regression model
}
\description{
This function fits a smoothed piecewise linear model to ln(abundance) data, with 
the assumption that abundances decrease linearly at first, but then hit a floor
and remain constant - consistent with a population that declines to the detection
limit of fluoresence.
}
