% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.gr.sat}
\alias{get.gr.sat}
\title{Extract exponential growth rate assuming exponential growth that saturates}
\usage{
get.gr.sat(x, y)
}
\arguments{
\item{x}{Time steps}

\item{y}{ln(abundance)}
}
\value{
This function returns a nonlinear least-squares regression model
}
\description{
This function fits a smoothed piecewise linear model to ln(abundance) data, with 
the assumption that abundances increase linearly at first, but then saturate and
remain constant.
}
