% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.growth.rate}
\alias{get.growth.rate}
\title{Extract exponential growth rate from a time series of ln(population abundance)}
\usage{
get.growth.rate(x, y, id, plot.best.Q = F, fpath = NA,
  methods = c("linear", "lag", "sat", "lagsat"), internal.r2.cutoff = 0,
  verbose = FALSE)
}
\arguments{
\item{x}{Time steps}

\item{y}{ln(abundance)}

\item{id}{Label corresponding to the population/strain/species of interest; used to determine the title and file name of saved plot, if any.}

\item{plot.best.Q}{logical; should the best fitting model be plotted?}

\item{fpath}{character; if best model is to be plotted, provide the file path for saving the plot}

\item{methods}{Must be a character vector containing one or more of \code{'linear'}, \code{'lag'}, \code{'sat'}, or \code{'lagsat'}}

\item{internal.r2.cutoff}{control parameter specifying the R2 criteria that may be applied to drop fits where the number of observations in the exponential portion is equal to 3. The default value of zero permits all fits of 3 obs to be considered.}
}
\value{
A data frame containing the identity of the best model, the content of the best model, the estimated slopes of the increasing linear portion of the regressions (ie, exponential growth rate), the standard errors associated with these slopes, and the full list of all models fit.
}
\description{
This meta-function takes a time series of abundance, and attempts to extract an 
estimate of exponential growth rate, using one or more of a suite of possible methods.
These methods allow for the possibility that exponential growth may lag or saturate,
or both, over the course of the time series. All selected methods are used to fit 
models to the time series. Subsequently, model comparison (based on AIC) is used to 
determine which model best fits the focal data.
}
\examples{

}

