% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_tools.R
\name{get.monod2.npc}
\alias{get.monod2.npc}
\title{Fit Monod curve with loss term to growth rate vs. nutrient concentration data}
\usage{
get.monod2.npc(nutrients, mu, method = "mle2", ...)
}
\arguments{
\item{nutrients}{Nutrient concentration}

\item{mu}{Exponential growth rate}

\item{method}{Specify which fitting algorithm to use, 'mle2' or 'grid.mle2'}

\item{...}{Additional arguments passed to grid.mle2 (e.g., control=list(maxit=2000))}
}
\description{
Note: this function currently does not use grid.mle2, as the regressions are usually pretty stable; add this 
feature later.
}
\examples{
(write some...)

}
