% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{get.nbcurve.tpc}
\alias{get.nbcurve.tpc}
\title{Fit Norberg curve to growth rate vs. temperature data}
\usage{
get.nbcurve.tpc(temp, mu, method = "grid.mle2", plotQ = F, conf.bandQ = T,
  fpath = NA, id = NA)
}
\arguments{
\item{temp}{Temperature}

\item{mu}{Exponential growth rate}

\item{method}{Specify which fitting algorithm to use, 'mle2' or 'grid.mle2'}

\item{plotQ}{Should regression be visualized?}

\item{fpath}{If visual requested, and valid file path provided here, plot will be saved as a .pdf file. Default is NA.}
}
\description{
Fit Norberg curve to growth rate vs. temperature data
}

