% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{get.R2}
\alias{get.R2}
\title{Calculate R2}
\usage{
get.R2(pds, obs)
}
\arguments{
\item{pds}{Predicted y values (could arise from predict(model) for an mle2 model)}

\item{obs}{Observed y values}
}
\value{
R2 value
}
\description{
1-sum((obs-pds)^2)/sum((obs-mean(obs))^2)
}
