% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{get.R2}
\alias{get.R2}
\title{Calculate R2 for an mle2 model}
\usage{
get.R2(model, obs)
}
\arguments{
\item{model}{An mle2 model object}

\item{obs}{Observed y values}
}
\value{
R2 value
}
\description{
Assuming a normal error distribution and an mle2 fit that uses the 'formula' syntax
(this function will not work if mle2 is handed a black-boxed NLL function)
}
