% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools.R
\name{get.topt}
\alias{get.topt}
\title{Get true optimum temperature from nbcurve()}
\usage{
get.topt(topt, w, b)
}
\arguments{
\item{topt}{Competitive optimum temperature}

\item{w}{Thermal niche width}

\item{b}{Exponential scaling}
}
\value{
Optimum temperature
}
\description{
Given the competitive optimum temperature and other Norberg parameters, calculate the
true optimum temperature.
}
