% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{glance.growth_rate_result}
\alias{glance.growth_rate_result}
\title{Glance method for suite of successful growth rate results}
\usage{
glance.growth_rate_result(object, ...)
}
\arguments{
\item{object}{Object of class growth_rate_result}

\item{\dots}{Additional arguments (not used)}
}
\description{
Glance method for suite of successful growth rate results
}
