% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthTools.R
\docType{package}
\name{growthTools}
\alias{growthTools-package}
\alias{growthTools}
\title{growthTools: A package for computing microbial growth rates and traits.}
\description{
The growthTools package does X, Y, and Z. (see: http://r-pkgs.had.co.nz/man.html#man-packages)
}
\section{get.gr functions}{

The get.gr family of functions does ...
}

