% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{linear_model}
\alias{linear_model}
\title{Defining growth model classes}
\usage{
linear_model()
}
\description{
Each growth model exists as both a 'growth_model' object and carries a specific
class determining which type of growth model it is. These functions establish
these class definitions.
}
