% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_tools.R
\name{monod_curve}
\alias{monod_curve}
\title{Monod curve equation}
\usage{
monod_curve(x, umax, k, z, log.k = FALSE, log.umax = FALSE)
}
\arguments{
\item{x}{nutrients}

\item{umax}{maximum growth rate}

\item{k}{half saturation constant}

\item{z}{intercept}

\item{log.k}{logical, is k specified on natural log scale}

\item{log.umax}{logical, is umax specified on natural log scale}
}
\value{
Predicted exponential growth rate at nutrient concentration x
}
\description{
Details here.
}
