% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_tools.R
\name{monod_curve}
\alias{monod_curve}
\title{Monod curve equation}
\usage{
monod_curve(x, umax, k, z = 0)
}
\arguments{
\item{x}{nutrients}

\item{umax}{maximum growth rate}

\item{k}{half saturation constant}

\item{z}{intercept, defaults to 0}
}
\value{
Predicted exponential growth rate at nutrient concentration x
}
\description{
Details here.
}
