% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{new_growth_fit}
\alias{new_growth_fit}
\title{Growth fit class creation}
\usage{
new_growth_fit(
  model,
  fit,
  data,
  results,
  status = "ok",
  error = NA,
  growth.rate.valid = FALSE
)
}
\arguments{
\item{model}{Growth model type}

\item{fit}{Actual fit of growth model}

\item{data}{List containing the x (time) and y (ln(abundance) data used for fitting}

\item{results}{Growth model results, including diagnostics/metrics from evaluating fit}

\item{status}{Status of growth model fit and evaluation}

\item{error}{Error(s) arising when model fit or evaluation failed}

\item{growth.rate.valid}{Does the fit meet conditions for robust growth rate estimate?}
}
\description{
Growth fit class creation
}
