% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{new_growth_fit}
\alias{new_growth_fit}
\title{Growth fit class creation}
\usage{
new_growth_fit(model, fit, metrics, growth.rate.valid = FALSE)
}
\arguments{
\item{model}{Growth model type}

\item{fit}{Actual fit of growth model}

\item{metrics}{Growth model diagnostics/metrics from evaluating fit}

\item{growth.rate.valid}{Does the fit meet conditions for robust growth rate estimate?}
}
\description{
Growth fit class creation
}
