% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{plot.growth_rate_result}
\alias{plot.growth_rate_result}
\title{Plot method for best model in the suite of growth rate results}
\usage{
\method{plot}{growth_rate_result}(object, main = NULL, ...)
}
\arguments{
\item{object}{Object of class growth_rate_result}

\item{main}{Plot title; defaults to name of best model}

\item{\dots}{Additional arguments (not used)}
}
\description{
Plot method for best model in the suite of growth rate results
}
