% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools_S3.R
\name{plot.tpc}
\alias{plot.tpc}
\title{Plotting function for single thermal performance curve}
\usage{
\method{plot}{tpc}(
  x,
  plot_ci = TRUE,
  plot_obs = TRUE,
  xlim = c(-2, 40),
  ylim = c(-0.2, 5),
  main = NA,
  fpath = NA,
  ...
)
}
\arguments{
\item{x}{The result of a single TPC curve fit from get.nbcurve.tpc or similar}

\item{plot_ci}{logical, should the resulting plot include 95\% confidence bands}

\item{plot_obs}{logical, should resulting plot include raw data}

\item{xlim}{x-axis range (temperature)}

\item{ylim}{y-axis range (adjusts internally from -0.2 to slightly above umax+CI)}

\item{main}{Character string providing plot title (usually id info for the plotted curve)}

\item{fpath}{If visual requested, and valid file path provided, save plot as a .pdf file}

\item{\dots}{Additional arguments (not used)}
}
\description{
Plotting function for single thermal performance curve
}
