% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{predict.growth_fit}
\alias{predict.growth_fit}
\title{Generic predict method for growth_fit object}
\usage{
\method{predict}{growth_fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class growth_fit}

\item{newdata}{New data (if any) to use for making predictions}

\item{\dots}{Additional arguments (not used)}
}
\description{
Generic predict method for growth_fit object
}
