% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCtools_S3.R
\name{predict.tpc}
\alias{predict.tpc}
\title{Predict values from a TPC fit}
\usage{
\method{predict}{tpc}(fit, newdata, se.fit = FALSE)
}
\arguments{
\item{fit}{The result of a single TPC curve fit from get.nbcurve.tpc}

\item{newdata}{A new data frame, containing a sequence of `temperature` values at which model predictions should be made; defaults to (-2,40)}

\item{se.fit}{logical; should standard error values be returned?}
}
\description{
Predict values from a TPC fit
}
