% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_S3.R
\name{predict_model}
\alias{predict_model}
\alias{predict_model.linear}
\alias{predict_model.lag}
\alias{predict_model.sat}
\alias{predict_model.lagsat}
\alias{predict_model.flr}
\alias{predict_model.satdecay}
\alias{predict_model.satdecay_ode_model}
\title{Prediction methods for growth models}
\usage{
predict_model(model, fit, newdata, ...)

\method{predict_model}{linear}(model, fit, newdata, ...)

\method{predict_model}{lag}(model, fit, newdata, ...)

\method{predict_model}{sat}(model, fit, newdata, ...)

\method{predict_model}{lagsat}(model, fit, newdata, ...)

\method{predict_model}{flr}(model, fit, newdata, ...)

\method{predict_model}{satdecay}(model, fit, newdata, ...)

\method{predict_model}{satdecay_ode_model}(model, fit, newdata, ...)
}
\arguments{
\item{model}{Growth model type}

\item{fit}{Fit of particular growth model type to the data, from fit_model}

\item{newdata}{New data (if any) used for making predictions; must contain column named 'x' for time values}

\item{\dots}{Additional arguments passed to fitting function (not used?)}
}
\description{
A suite of growth models is available which can be fit to time series of 
ln(abundance) over time and used to estimate population exponential growth
rate. The generic method for generating predictions from these growth_fit
objects, predict_model(), dispatches to growth model specific prediction
functions. All of these are ultimately accessed through predict.growth_fit().
}
