% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_tools.R, R/nutrient_tools.R
\name{print.npc}
\alias{print.npc}
\title{Print method for NPC objects}
\usage{
\method{print}{npc}(x, ...)

\method{print}{npc}(x, ...)
}
\arguments{
\item{x}{Object of class npc}

\item{\dots}{Additional arguments (not used)}
}
\description{
Print method for NPC objects

Print method for NPC objects
}
