% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{satdecay.ode}
\alias{satdecay.ode}
\title{ODE Equations for modeling abundance time series}
\usage{
satdecay.ode(x, alpha, vmax, cpar, dpar, r0, n0)
}
\arguments{
\item{x}{Time variable}

\item{alpha}{Affinity; defined as vmax over k, must be >0}

\item{vmax}{Maximum uptake rate, must be >0}

\item{cpar}{Proportion of dead biomass returned to the labile resource pool, must be 0 <= cpar <= 1}

\item{dpar}{Mortality rate, must be >0}

\item{r0}{Initial resource concentration at time t=0 (unitless, scaled by k, must be >0)}

\item{n0}{Initial log(abundance) at time t=0}
}
\value{
log(abundance) at time x as a function of model parameters
}
\description{
Intended to allow the extraction of exponential growth rates from time series 
while accounting for unobserved, resource-dependent growth, saturating abundance,
and potentially declines in abundance after saturation due to accumulation of
resource in a recalcitrant pool. The satdecay() function is a piecewise linear
approximation of these dynamics.
}
\details{
NOTE: currently, this function depends on computation of numerical ODE solution
in Julia (for computational efficiency), and will not work without a functional
connection to Julia.
}
