% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sqfunc}
\alias{sqfunc}
\title{Helper function for smoothed lagged/saturating abundance equations}
\usage{
sqfunc(x, b, s)
}
\arguments{
\item{x}{time}

\item{b}{slope}

\item{s}{smoothing parameter}
}
\description{
Helper function for smoothed lagged/saturating abundance equations
}
